+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+* Column instances now have a constraints list that is applied to its definition.
+
+### Changed
+* Renamed column method "statement" to "definition" to better fit SQLite terminology.
+
 ## [0.1.1] - 2026-05-27
 
 ### Added
@@ -43,5 +51,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     * pypi-publish.yml
     * pypi-test.yml
 
+[Unreleased]: https://github.com/MICHI64N/sqlitables/compare/v0.1.1...HEAD
 [0.1.1]: https://github.com/MICHI64N/sqlitables/compare/v0.1.0...v0.1.1
 [0.1.0]: https://github.com/MICHI64N/sqlitables/releases/tag/v0.1.0